meu_projeto/requirements.txt
flask
flask-sqlalchemy
flask-migrate
pymysql
cryptography
flask-login
flask-bcrypt
python-dotenv
pytz
gunicorn